refactor: split sandbox resource builders
