+++ services/api-rs/crates/centaur-sandbox-agent-k8s/src/resources/sandbox/env.rs
@@ -1,0 +1,133 @@
+use std::collections::BTreeMap;
+
+use centaur_sandbox_core::SandboxSpec;
+use k8s_openapi::api::core::v1::EnvVar;
+
+use super::super::common::env_var;
+use super::super::iron_proxy::ResolvedIronProxy;
+
+pub(super) fn env_vars(
+    spec: &SandboxSpec,
+    resolved_iron_proxy: Option<&ResolvedIronProxy>,
+) -> Option<Vec<EnvVar>> {
+    let mut env = BTreeMap::<String, String>::new();
+    for item in &spec.env {
+        env.insert(item.name.clone(), item.value.clone());
+    }
+    if let Some(resolved_iron_proxy) = resolved_iron_proxy {
+        for (name, value) in &resolved_iron_proxy.placeholder_env {
+            env.entry(name.clone()).or_insert_with(|| value.clone());
+        }
+        for (name, value) in &resolved_iron_proxy.pg_dsn_env {
+            env.entry(name.clone()).or_insert_with(|| value.clone());
+        }
+        let api_host = env
+            .get("CENTAUR_API_URL")
+            .and_then(|value| host_from_url(value));
+        let no_proxy_extra = ["NO_PROXY", "no_proxy"]
+            .into_iter()
+            .filter_map(|name| env.get(name).map(String::as_str))
+            .collect::<Vec<_>>();
+        for (name, value) in proxy_env(
+            &resolved_iron_proxy.proxy_host,
+            resolved_iron_proxy.proxy_port,
+            api_host.as_deref(),
+            &no_proxy_extra,
+        ) {
+            env.insert(name, value);
+        }
+    }
+    (!env.is_empty()).then(|| {
+        env.into_iter()
+            .map(|(name, value)| env_var(&name, &value))
+            .collect()
+    })
+}
+
+fn proxy_env(
+    proxy_host: &str,
+    proxy_port: u16,
+    api_host: Option<&str>,
+    no_proxy_extra: &[&str],
+) -> BTreeMap<String, String> {
+    let proxy_url = format!("http://{proxy_host}:{proxy_port}");
+    let no_proxy = no_proxy_value(proxy_host, api_host, no_proxy_extra);
+    BTreeMap::from([
+        ("FIREWALL_HOST".to_owned(), proxy_host.to_owned()),
+        ("FIREWALL_PROXY_PORT".to_owned(), proxy_port.to_string()),
+        ("HTTP_PROXY".to_owned(), proxy_url.clone()),
+        ("HTTPS_PROXY".to_owned(), proxy_url.clone()),
+        ("http_proxy".to_owned(), proxy_url.clone()),
+        ("https_proxy".to_owned(), proxy_url),
+        ("NO_PROXY".to_owned(), no_proxy.clone()),
+        ("no_proxy".to_owned(), no_proxy),
+        (
+            "NODE_EXTRA_CA_CERTS".to_owned(),
+            "/firewall-certs/ca-cert.pem".to_owned(),
+        ),
+        (
+            "REQUESTS_CA_BUNDLE".to_owned(),
+            "/firewall-certs/ca-cert.pem".to_owned(),
+        ),
+        (
+            "CURL_CA_BUNDLE".to_owned(),
+            "/firewall-certs/ca-cert.pem".to_owned(),
+        ),
+        (
+            "SSL_CERT_FILE".to_owned(),
+            "/firewall-certs/ca-cert.pem".to_owned(),
+        ),
+        (
+            "GIT_SSL_CAINFO".to_owned(),
+            "/firewall-certs/ca-cert.pem".to_owned(),
+        ),
+    ])
+}
+
+fn no_proxy_value(proxy_host: &str, api_host: Option<&str>, extra_values: &[&str]) -> String {
+    let mut hosts = vec![
+        "localhost".to_owned(),
+        "127.0.0.1".to_owned(),
+        "::1".to_owned(),
+        proxy_host.to_owned(),
+        "api".to_owned(),
+        "victoriametrics".to_owned(),
+        "victorialogs".to_owned(),
+    ];
+    if let Some(api_host) = api_host.filter(|value| !value.is_empty()) {
+        hosts.push(api_host.to_owned());
+    }
+    for value in extra_values {
+        hosts.extend(
+            value
+                .split(',')
+                .map(str::trim)
+                .filter(|host| !host.is_empty())
+                .map(ToOwned::to_owned),
+        );
+    }
+    let mut deduped = Vec::new();
+    for host in hosts {
+        if !deduped.contains(&host) {
+            deduped.push(host);
+        }
+    }
+    deduped.join(",")
+}
+
+fn host_from_url(value: &str) -> Option<String> {
+    let value = value.trim();
+    let without_scheme = value
+        .split_once("://")
+        .map(|(_, rest)| rest)
+        .unwrap_or(value);
+    let authority = without_scheme.split('/').next()?.trim();
+    let host_port = authority
+        .rsplit_once('@')
+        .map(|(_, host_port)| host_port)
+        .unwrap_or(authority);
+    let host = host_port
+        .split_once(':')
+        .map_or(host_port, |(host, _)| host);
+    (!host.is_empty()).then(|| host.to_owned())
+}
