+++ services/api-rs/crates/centaur-sandbox-agent-k8s/src/resources/sandbox/status.rs
@@ -1,0 +1,41 @@
+use centaur_sandbox_core::SandboxStatus;
+use k8s_openapi::api::core::v1::Pod;
+
+pub(crate) fn sandbox_status_from_pod(replicas: i32, pod: Option<&Pod>) -> SandboxStatus {
+    if replicas == 0 {
+        return SandboxStatus::Suspended;
+    }
+    // The backing Pod Ready condition is the attach boundary; phase alone can be Running while
+    // the sandbox is still not ready for I/O.
+    let Some(pod) = pod else {
+        return SandboxStatus::Created;
+    };
+    if pod.metadata.deletion_timestamp.is_some() {
+        return SandboxStatus::Created;
+    }
+
+    let phase = pod
+        .status
+        .as_ref()
+        .and_then(|status| status.phase.as_deref())
+        .unwrap_or("unknown")
+        .to_ascii_lowercase();
+    match phase.as_str() {
+        "running" if pod_ready(pod) => SandboxStatus::Running,
+        "running" | "pending" => SandboxStatus::Created,
+        "succeeded" | "failed" => SandboxStatus::Stopped,
+        "unknown" => SandboxStatus::Unknown("unknown".to_owned()),
+        other => SandboxStatus::Unknown(other.to_owned()),
+    }
+}
+
+fn pod_ready(pod: &Pod) -> bool {
+    pod.status
+        .as_ref()
+        .and_then(|status| status.conditions.as_ref())
+        .is_some_and(|conditions| {
+            conditions
+                .iter()
+                .any(|condition| condition.type_ == "Ready" && condition.status == "True")
+        })
+}
